Fix CI after Heroku-24 default stack change
After the change in default stack to Heroku-24, tests were failing with:

```
.WARNING: dyno state is "crashed"
```

This is due to the test app having a `web` process that uses Python, when the test app doesn't configure the Python buildpack.

It turns out the test app doesn't even need the web process or Python - it appears to be a copy-paste leftover from the heroku-exec tests (on which these tests were based).

As such, I've removed the unused parts of the test, which also fixes the failures on Heroku-24.